Updated Payment Controller to allow for Unit Testing Created 2 unit tests.
